+++ pakyow-core/lib/pakyow/environment.rb
@@ -19,6 +19,7 @@ require "pakyow/behavior/erroring"
 require "pakyow/behavior/initializers"
 require "pakyow/behavior/input_parsing"
 require "pakyow/behavior/plugins"
+require "pakyow/behavior/rescuing"
 require "pakyow/behavior/silencing"
 require "pakyow/behavior/tasks"
 require "pakyow/behavior/timezone"
@@ -274,6 +275,7 @@ module Pakyow
   include Behavior::Initializers
   include Behavior::InputParsing
   include Behavior::Plugins
+  include Behavior::Rescuing
   include Behavior::Silencing
   include Behavior::Tasks
   include Behavior::Timezone
@@ -318,10 +320,6 @@ module Pakyow
     # Name of the environment
     #
     attr_reader :env
-
-    # Any error encountered during the boot process
-    #
-    attr_reader :error
 
     # Global log output.
     #
@@ -400,6 +398,10 @@ module Pakyow
 
         @__loaded = true
       end
+    rescue ApplicationError => error
+      raise error
+    rescue ScriptError, StandardError => error
+      raise EnvironmentError.build(error)
     end
 
     # Returns true if the environment has loaded.
@@ -464,6 +466,10 @@ module Pakyow
       end
 
       self
+    rescue ApplicationError => error
+      raise error
+    rescue ScriptError, StandardError => error
+      raise EnvironmentError.build(error)
     end
 
     # Returns true if the environment has been setup.
@@ -493,6 +499,10 @@ module Pakyow
       end
 
       self
+    rescue ApplicationError => error
+      raise error
+    rescue ScriptError, StandardError => error
+      raise EnvironmentError.build(error)
     end
 
     # Returns true if the environment has booted.